pipeline/pyproject.toml
[project]
name = "gpu-cost-pipeline"
version = "0.1.0"
description = "Data pipeline for GPU Cost Explorer"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
# Full pipeline dependencies (model enrichment + GPU pricing)
pipeline = [
    "beautifulsoup4>=4.12",
    "gpuhunt>=0.1",
    "httpx>=0.27",
    "pydantic>=2.7",
]

# Development tools
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "pre-commit>=3.7",
]

# All dependencies (pipeline + dev)
all = [
    "gpu-cost-pipeline[pipeline,dev]",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]
